requirements.txt
arrayqueues==1.3.1
split_dataset
lightparam>=0.4, <0.5
pyqt5>=5.15
pyzmq
QDarkStyle==2.7
markdown==3.2.2
click
traitlets
nidaqmx
toml
scopecuisine>=0.1.2
pyqtgraph
napari
numba==0.52.0
scikit-image
llvmlite==0.35.0
numpy==1.19.5
